+++ RELEASE_NOTES.md
@@ -1,3 +1,5 @@
+- [Vulkan] Fixed swapchain creation for Wayland
+
 ### 5.6.0
 - https://github.com/aardvark-platform/aardvark.rendering/wiki/Aardvark-Rendering-5.6-changelog